+++ src/test/pretty/path-type-bounds.rs
@@ -14,13 +14,12 @@
 trait Tr {
     fn dummy(&self) { }
 }
-impl Tr for int { }
+impl Tr for isize { }
 
 fn foo<'a>(x: Box<Tr+ Sync + 'a>) -> Box<Tr+ Sync + 'a> { x }
 
 fn main() {
     let x: Box<Tr+ Sync>;
 
-    Box::new(1) as Box<Tr+ Sync>;
+    Box::new(1isize) as Box<Tr+ Sync>;
 }
-
