+++ src/test/run-pass/associated-types-issue-20371.rs
@@ -11,8 +11,12 @@
 // Test that we are able to have an impl that defines an associated type
 // before the actual trait.
 
+// pretty-expanded FIXME #23616
+
+#![feature(core)]
+
 use std::marker::MarkerTrait;
 
-impl X for f64 { type Y = int; }
+impl X for f64 { type Y = isize; }
 trait X : MarkerTrait { type Y; }
 fn main() {}
