+++ src/test/run-pass/block-fn-coerce.rs
@@ -8,10 +8,12 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-fn force<F>(f: F) -> int where F: FnOnce() -> int { return f(); }
+// pretty-expanded FIXME #23616
+
+fn force<F>(f: F) -> isize where F: FnOnce() -> isize { return f(); }
 
 pub fn main() {
-    fn f() -> int { return 7; }
+    fn f() -> isize { return 7; }
     assert_eq!(force(f), 7);
     let g = {||force(f)};
     assert_eq!(g(), 7);
