+++ src/test/run-pass/borrowed-ptr-pattern-3.rs
@@ -8,7 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-fn foo<'r>(s: &'r uint) -> bool {
+// pretty-expanded FIXME #23616
+
+fn foo<'r>(s: &'r usize) -> bool {
     match s {
         &3 => true,
         _ => false
