+++ src/test/run-pass/const-enum-vec-ptr.rs
@@ -8,7 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-enum E { V1(int), V0 }
+// pretty-expanded FIXME #23616
+
+enum E { V1(isize), V0 }
 static C: &'static [E] = &[E::V0, E::V1(0xDEADBEE), E::V0];
 
 pub fn main() {
