+++ src/test/run-pass/const-expr-in-fixed-length-vec.rs
@@ -11,9 +11,11 @@
 // Check that constant expressions can be used for declaring the
 // type of a fixed length vector.
 
+// pretty-expanded FIXME #23616
+
 pub fn main() {
 
-    const FOO: uint = 2;
-    let _v: [int; FOO*3];
+    const FOO: usize = 2;
+    let _v: [isize; FOO*3];
 
 }
