+++ src/test/run-pass/const-expr-in-vec-repeat.rs
@@ -10,9 +10,11 @@
 
 // Check that constant expressions can be used in vec repeat syntax.
 
+// pretty-expanded FIXME #23616
+
 pub fn main() {
 
-    const FOO: uint = 2;
+    const FOO: usize = 2;
     let _v = [0; FOO*3*2/2];
 
 }
