+++ src/test/run-pass/const-negative.rs
@@ -10,7 +10,9 @@
 
 // Issue #358
 
-static toplevel_mod: int = -1;
+// pretty-expanded FIXME #23616
+
+static toplevel_mod: isize = -1;
 
 pub fn main() {
     assert_eq!(toplevel_mod, -1);
